requirements.txt
fastapi[all]~=0.104.0
torch~=2.1.0
torchvision~=0.16.0
requests~=2.31.0
Pillow~=10.1.0
redis~=5.0.1
rq~=1.15.1
python-multipart
matplotlib~=3.8.2
opencv-python~=4.8.1.78
numpy~=1.26.1
python-magic~=0.4.27